Feat: Add support so that the nav and status bar can be transparent
The nav and status is set to be transparent. Therefore the window
has been set to take up the size of the screen. The content of the
app starts from the bottom of the screen which overlaps with the
navigation bar, so support was add to:
1) get the size of the navigation bar
2) detect where the navigation bar is depending on the orientation
and device type (its always at the bottom on tablets)
3) add nav bars height to the side of where the nav bar is located